Pushed the player controls to a separate file.
